+++ crates/telescope/src/app/windows/about.rs
@@ -7,7 +7,8 @@ use eframe::egui::Vec2;
 impl TelescopeApp {
     #[tracing::instrument(skip(self, ctx))]
     pub(crate) fn open_about_window(&mut self, ctx: &egui::Context) {
-        egui::Window::new("About Telescope")
+        egui::Window::new(t!("about.title"))
+            .id(egui::Id::new("about_window"))
             .fixed_size((400.0, 200.0))
             .open(&mut self.open[0])
             .show(ctx, |ui| {
@@ -20,8 +21,8 @@ impl TelescopeApp {
                         ui.add_space(10.0);
                         ui.heading("Telescope");
                         ui.strong("v ".to_owned() + env!("CARGO_PKG_VERSION"));
-                        ui.label("Author: Rafael Amador");
-                        ui.label("Licensed under MIT");
+                        ui.label(t!("about.author", name = "Rafael Amador"));
+                        ui.label(t!("about.license", license = "MIT"));
                         if ui.link("https://github.com/rafaga/telescope").clicked() {
                             let _a = open::that("https://github.com/rafaga/telescope");
                         }